fixes #918
